# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: API для управления пользователями, картами и переводами
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/cards:
    get:
      tags:
        - Cards API
      summary: Получить список карт
      description: ADMIN видит все карты, USER может видеть только свои
      operationId: getCards
      parameters:
        - name: status
          in: query
          description: Статус карты
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
        - name: username
          in: query
          description: Имя пользователя
          required: false
          schema:
            type: string
          example: john_doe1999
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
        - name: sort
          in: query
          description: "Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported."
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Страница карт получена
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PageCardInfoResponse"
        "401":
          description: Неавторизован
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
        - Cards API
      summary: Создать карту
      description: Только ADMIN может создать карту
      operationId: createCard
      requestBody:
        description: Запрос на создание карты
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
        required: true
      responses:
        "201":
          description: Карта создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardInfoResponse"
        "400":
          description: Неверные данные владельца
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Неавторизован
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Пользователь не найден
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/auth:
    post:
      tags:
        - Auth API
      summary: Аутентифицироваться
      description: Публичный эндпоинт
      operationId: authenticate
      requestBody:
        description: Пара логин/пароль
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Аутентификация прошла успешно
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Валидация пароля или username не прошла
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Пароль или username не подходят
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка генерации JSON Web Token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/cards/{cardId}:
    get:
      tags:
        - Cards API
      summary: Получить карту по ID
      description: ADMIN или владелец карты может получить данные карты
      operationId: getCard
      parameters:
        - name: cardId
          in: path
          description: Идентификатор карты
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardInfoResponse"
        "401":
          description: Неавторизован
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Карта не найдена
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка расшифровки карты
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      tags:
        - Cards API
      summary: Удалить карту
      description: Только ADMIN может удалить карту
      operationId: deleteCard
      parameters:
        - name: cardId
          in: path
          description: Идентификатор карты
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Карта удалена
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Неавторизован
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Карта не найдена
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    patch:
      tags:
        - Cards API
      summary: Обновить статус карты
      description: ADMIN или владелец карты может обновить статус
      operationId: updateCardStatus
      parameters:
        - name: cardId
          in: path
          description: Идентификатор карты
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Запрос на обновление статуса карты
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardStatusRequest"
        required: true
      responses:
        "200":
          description: Статус карты обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardInfoResponse"
        "400":
          description: Ошибка смены статуса карты
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Неавторизован
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Карта не найдена
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
...
    CreateCardRequest:
      type: object
      description: Запрос на создание карты
      properties:
        userId:
          type: string
          format: uuid
          description: Идентификатор пользователя
          example: 64e387b5-f43d-4595-85c8-51cef15e647a
        ownerName:
          type: string
          description: Имя владельца
          example: IVAN IVANOV
          maxLength: 32
          minLength: 5
      required:
        - ownerName
        - userId
    CardInfoResponse:
      type: object
      description: Информация о карте
      properties:
        cardId:
          type: string
          format: uuid
          description: Идентификатор карты
          example: 64e387b5-f43d-4595-85c8-51cef15e647a
        maskedNumber:
          type: string
          description: Замаскированный PAN
          example: "**** **** **** 1234"
        cardStatus:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        ownerName:
          type: string
          description: Имя владельца
          example: IVAN IVANOV
        ownerId:
          type: string
          format: uuid
          description: Идентификатор владельца
          example: 64e387b5-f43d-4595-85c8-51cef15e647a
        expirationDate:
          type: string
          description: ГОД-МЕСЯЦ окончания срока действия карты
          example: 2026-04
        balance:
          type: number
          description: Баланс карты
          example: "100.00"
    LoginRequest:
      type: object
      description: Запрос на вход в учётную запись
      properties:
        username:
          type: string
          description: Username пользователя
          example: john_doe123
        password:
          type: string
          description: Пароль пользователя
          example: v3ry_s7r0n9_p4$$w0rd
      required:
        - password
        - username
    LoginResponse:
      type: object
      description: Ответ на запрос на вход в учётную запись
      properties:
        token:
          type: string
          description: JSON Web Token
    UpdateCardStatusRequest:
      type: object
      description: Запрос на обновление статуса карты
      properties:
        cardStatus:
          type: string
          description: Новый статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
      required:
        - cardStatus
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageCardInfoResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardInfoResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT